Add signals to Net_Connection interface
Make it possible for Net_Connection to be used in an event-driven
fashion with the messagesReady() signal, which is emitted each time one
or more new messages arrive.

The disconnected() signal is emitted when the connection enters an error
state or has been closed.

For now the ->Run() method is still available so that the client and
server can be converted to an event-driven architecture independently.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>