Keep a single QNetworkAccessManager in qtclient
The ServerBrowser uses QNetworkAccessManager for HTTP requests.
Applications should use a single QNetworkAccessManager instance so
cookies and other state are shared between requests.

This patch moves QNetworkAccessManager out of ServerBrowser and gives it
a member variable in MainWindow.

The ConnectDialog and ServerBrowser constructors need to take the
QNetworkAccessManager as an argument.  This keeps the classes decoupled
and makes them easy to test or reusable.

Note that the comment in ServerBrowser about deleting reply objects is
outdated - we already connect the finish() signal to
reply->deleteLater(). It is safe to drop the comment, the reply object
will be freed.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>