Remove member used for temporary variable
Parameter was set to member and passed to slot.
That make class have a state, when loadServerList was called twice
before start completeDownloadServerList, they may get unexpected "reply"
instance. This change ensure the sender is the "reply" instance who invoked
the slot. ServerBrowser class will get close to a reentrant.

Signed-off-by: Ikkei Shimomura (tea) <Ikkei.Shimomura@gmail.com>